+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### Unreleased
 
 - Rename PluginInterface main method from `run()` to `__invoke()` 
+- Rename `addPlugin()` to `addAfterPlugin()`
 
 ### 1.2.0
 ### 2023-04-29